add stratified sampling by age
For RMSE calculations by age, don't want to bias results given uneven sampling by ages. This could lead to skewed outcomes when certain components of the growth curve are over/under represented in the data. This routine subsets 100 unique samples for each age bin (or the max # samples available, whichever is smaller) and then calculates the test statistic.

Note the by age results are only used for illustration and not in choosing the best model -- we decided not to construct a franken-prediction of mixed models & ages.